pyproject.toml
[project]
name = "thanatosagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.124.4",
    "httpx>=0.28.1",
    "langchain-groq>=1.1.1",
    "langgraph>=1.0.5",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
]
